Add token to the Provider secret check. (#6)
Update the token whenever a change is detected.
Remove the token from the providerSecret definition to make sure it is always set with the NEW value when creating the secret for the first time.

Note: The OnwerReference is still missing, to add this, we would likely need to re-order the Provider create step to be before the Provider secret create.

Signed-off-by: Joshua Packer <jpacker@redhat.com>